// src/interfaces/sprint_stats.ts
export interface SprintStats {
  executed: number; // К выполнению
  work: number; // В работе
  done: number; // Сделано
  remove: number; // Снято
  backlog: number; // Бэклог
  backlog1: number; // Бэклог1
  backlog2: number; // Бэклог2
  block: number; // Заблокировано задач в Ч/Д
}